chore: refactored service discovery logic to hit internal hrm endpoints
