+++ packages/service-discovery/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../tsconfig.packages-base.json",
+  "include": ["*.ts"],
+  "compilerOptions": {
+    "outDir": "./dist",
+  },
+}
